std::variant and other improvements

+++ test/xson/examples.cpp
@@ -73,5 +73,5 @@ TEST(Examples,Parse)
 
     string name = result["Name"s];
 
-    double number = result["Lucky Numbers"s][1];
+    int number = result["Lucky Numbers"s][1];
 }
